### 2018-09-01 #923 @whedon
[PRE REVIEW]: Embeddings.jl: easy access to pretrained word embeddings from julia
**Submitting author:** @oxinabox (<a href="http://orcid.org/0000-0003-1386-1646">Lyndon White</a>)
**Repository:** <a href="https://github.com/JuliaText/Embeddings.jl/" target ="_blank">https://github.com/JuliaText/Embeddings.jl/</a>
**Version:** v0.3.0
