### 2018-09-04 #923 @Kevin-Mattheus-Moerman
@weijianzhang is this submission of interest to you? Would you be able to help review it for [JOSS](http://joss.theoj.org/)? Thanks.
