### 2018-09-04 a5fbf0f
Update paper.md
