### 2018-09-04 e70d018
Update paper.md
