### 2018-09-10 #923 @AsafManela
I cannot, too busy these days. Sorry,

- Asaf

On Tue, Sep 4, 2018, 5:49 AM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @weijianzhang <https://github.com/weijianzhang> @aviks
> <https://github.com/aviks> @AsafManela <https://github.com/AsafManela>
> @hshindo <https://github.com/hshindo> is this submission of interest to
> you? Would you be able to help review it for the Journal of Open Source
> Software (JOSS) <http://joss.theoj.org/>?
>
> The review process
> <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#the-joss-review-process>
> focuses on the software 💾 <https://github.com/JuliaText/Embeddings.jl/>,
> and a short paper 📜
> <https://github.com/openjournals/joss-papers/blob/joss.00923/joss.00923/10.21105.joss.00923.pdf>
> .
>
> Let me know if you have questions. Thanks.
>
