### 2018-09-14 #923 @Kevin-Mattheus-Moerman
@weijianzhang @aviks @hshindo @johnmyleswhite @abelsiqueira @dpo @vchuravy is this submission of interest to you? Would you be able to help review it for  [the Journal of Open Source Software (JOSS)](http://joss.theoj.org/)?

The [review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#the-joss-review-process) focuses on [the software :floppy_disk:](https://github.com/JuliaText/Embeddings.jl/), and a [short paper :scroll:](https://github.com/openjournals/joss-papers/blob/joss.00923/joss.00923/10.21105.joss.00923.pdf).

Let me know if you have questions. Thanks.
