### 2018-09-14 #923 @abelsiqueira
Sorry, I have no expertise on the subject. Our NLP stands for Nonlinear Programming.
