### 2018-09-17 5a40305
Added ConceptNetNumberbatch initial support
