### 2018-09-17 66639c0
Another round of bugfixes
