### 2018-09-19 9d86441
Implemented review remarks, #14
