### 2018-09-27 #923 @xuedong
@Kevin-Mattheus-Moerman Hi, I only have a limited experience on NLP, I don't know if I'm well qualified for this review.
