### 2018-09-27 #923 @ssfrr
@Kevin-Mattheus-Moerman similarly I have machine learning experience but not NLP specifically. I'm happy to review for software/documentation quality I can't really speak to whether the included embeddings are the best choices for NLP.
