### 2018-09-27 f518338
Remove dead link to docs
