+++ README.md
@@ -1,7 +1,6 @@
 # Embeddings
 
-[![Stable](https://img.shields.io/badge/docs-stable-blue.svg)](https://JuliaText.github.io/Embeddings.jl/stable)
-[![Latest](https://img.shields.io/badge/docs-latest-blue.svg)](https://JuliaText.github.io/Embeddings.jl/latest)
+
 [![Build Status](https://travis-ci.org/JuliaText/Embeddings.jl.svg?branch=master)](https://travis-ci.org/JuliaText/Embeddings.jl)
 [![CodeCov](https://codecov.io/gh/JuliaText/Embeddings.jl/branch/master/graph/badge.svg)](https://codecov.io/gh/JuliaText/Embeddings.jl)
 
