### 2018-10-10 #1013 @whedon
[REVIEW]: Embeddings.jl: easy access to pretrained word embeddings from julia
**Submitting author:** @oxinabox (<a href="http://orcid.org/0000-0003-1386-1646">Lyndon White</a>)
**Repository:** <a href="https://github.com/JuliaText/Embeddings.jl/" target ="_blank">https://github.com/JuliaText/Embeddings.jl/</a>
**Version:** 0.3.1
