### 2019-02-26 #1013 @aviks
Everything looks good apart from a couple of documentation nits that I've raised on the repo.
