### 2019-02-27 d4ba004
Add install instructions, closes #20
