### 2019-03-01 #1013 @Kevin-Mattheus-Moerman
@oxinabox I see you have a short contributing section in your readme. Would you be okay with expanding your community guidelines (last unticked box above) a bit, e.g. using the steps/info below?

- [ ] Please add [community guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines) i.e. in the form of a `CONTRIBUTING.md` file (see also here https://help.github.com/articles/github-community-guidelines/) and a [code of conduct](https://opensource.guide/code-of-conduct/) (`COC.md`) document (good templates are available here: https://www.contributor-covenant.org/).

- [ ] I recommend that you link to the CONTRIBUTING and COC files in the README file.
