### 2019-03-01 #1013 @Kevin-Mattheus-Moerman
The other unticked box is installation instructions.
@oxinabox can you also expand your installation instructions. @aviks if you have more feedback on this to point @oxinabox in the right direction (i.e. what you feel is missing) that would be great. Thanks again for your review work.
