### 2019-03-01 0e851a7
Add more to contributing section.
