### 2019-03-01 #1013 @oxinabox
I added an Installation section the other day. https://github.com/JuliaText/Embeddings.jl/blob/master/README.md#installation
I do not think there are more instructions I can give.

I have further expanded the Contributing section, and added reference to the Julia Community Standards.
I'm not sure if moving this section of the document to a CONTRIBUTING.md would make it more, or less likely to be read.
