### 2019-03-13 #1013 @oxinabox
- **correct:** `https://doi.org/10.1137/141000671 may be missing for title: Julia: A Fresh Approach to Numerical Computing`
- **Incorrect:** `https://doi.org/10.1108/09504121211251880 may be missing for title: DataDeps.jl: Repeatable Data Setup for Replicable Data Science`
    - this paper has been stuck in some review hole for almost a year, so the citation is for the preprint. idk where this DOI is from.
- **correct:** `https://doi.org/10.1162/tacl_a_00051 may be missing for title: Enriching Word Vectors with Subword Information`
- **Incorrect:** `https://doi.org/10.1162/tacl_a_00051 may be missing for title: Learning Word Vectors for 157 Languages`
    - This is the same DOI as above so it can't be for this paper too.
- **Incorrect:** `https://doi.org/10.2172/1209543 may be missing for title: Efficient estimation of word representations in vector space`
- **correct:** `https://doi.org/10.3115/v1/d14-1162 may be missing for title: GloVe: Global Vectors for Word Representation`

I should update the bibliography to use these DOIs that are correct?
