### 2019-03-20 #1013 @Kevin-Mattheus-Moerman
@oxinabox any luck updaring the .bib file? Let me know if you need help or have questions.
