+++ paper/paper.bib
@@ -36,7 +36,7 @@ eprint = {https://doi.org/10.1137/141000671}
   Pages                    = {135--146},
   Volume                   = {5},
   Timestamp                = {2017.11.28},
-  Url                      = {https://arxiv.org/abs/1607.04606}
+  Url                      = {https://arxiv.org/abs/1607.04606},
   doi                      = {10.1162/tacl_a_00051}
 }
 
@@ -64,7 +64,7 @@ eprint = {https://doi.org/10.1137/141000671}
   Year                     = {2014},
   Pages                    = {1532--1543},
   Timestamp                = {2015.12.15},
-  Url                      = {http://www.aclweb.org/anthology/D14-1162}
+  Url                      = {http://www.aclweb.org/anthology/D14-1162},
   doi                      = {10.3115/v1/d14-1162}
 }
 
