### 2019-03-21 3d9dece
Fix double bracketted citations
