### 2019-03-21 #1013 @Kevin-Mattheus-Moerman
Some minor comments on the paper:

- [ ] Can you define the acronyms NLP, LSTM

- [ ] Either introduce NLP and use it consistently or remove it and say natural language processing throughout (it appears natural language processing is spelled out later in the text).

- [ ] Can you use the same form in the bullet list? By form I mean that the word embeddings follows after the bracketed references e.g.:
- Bla bla ProjectSoftWareName (REFERENCES) embeddings bla bla

- [ ] The 3rd bullet point contains two grouped reference sets, should these be merged?:  (Grave, Bojanowski, Gupta, Joulin, & Mikolov, 2018), (Bojanowski, Grave, Joulin, & Mikolov, 2017)

- [ ] Could you make sure the reference to "DataDeps.jl: Repeatable
Data Setup for Replicable Data Science" will also show up with a "Retrieved from" link? I believe it is https://arxiv.org/pdf/1808.01091.pdf.
