### 2019-03-25 f950b10
clean up based on feedback
