+++ paper/paper.bib
@@ -25,6 +25,7 @@ eprint = {https://doi.org/10.1137/141000671}
   keywords      = {Computer Science - Software Engineering},
   primaryclass  = {cs.SE},
   timestamp     = {2018.08.09},
+  url           = {https://arxiv.org/pdf/1808.01091.pdf},
 }
 
 @Article{fasttext,
