### 2019-03-30 7c50b74
Add basic example of use
