### 2019-04-08 #1013 @Kevin-Mattheus-Moerman
@oxinabox looks good.

One minor thing. Currently it says "an long", this should be "a long". Also, since now you do not reuse the acronym LSTM, consider removing it all together, or if the acronym is used more often than the spelled out form use:  "LSTM (long short term memory)" rather than "long short term memory (LSTM)".
