### 2019-04-08 f12e909
Put LSTM abrv first
