### 2019-04-09 #1013 @Kevin-Mattheus-Moerman
Great. Thanks @oxinabox. At this point can you provide a DOI to an archived version of the final reviewed software through a service like [ZENODO](https://zenodo.org/) (some find [these steps](https://guides.github.com/activities/citable-code/) useful)? Once you've done that can you provide the version tag of the final version of the reviewed software?
