### 2019-04-09 #1013 @Kevin-Mattheus-Moerman
@openjournals/joss-eics I recommend this work is accepted in JOSS
